// src/wrappers/main-wrapper/service/user/vdc/dto/get-vdc-compute-policy.dto.ts
export interface GetVdcComputePolicy {
  resultTotal: number;
  pageCount: number;
  page: number;
  pageSize: number;
  associations: null;
  values: Value[];
}

export interface Value {
  id: string;
  description: string;
  policyType: string;
  name: string;
  pvdcComputePolicy: null; //unknown
  compatibleVdcTypes: null; //unknown
  isSizingOnly: boolean;
  hasSizingInfo: boolean;
  pvdcId: null; //unknown
  isAutoGenerated: boolean;
  vgpuProfiles: null;
  pvdcVgpuClustersMap: null;
  cpuSpeed: null; //unknown
  memory: null; //unknown
  cpuCount: null; //unknown
  coresPerSocket: null; //unknown
  memoryReservationGuarantee: null; //unknown
  cpuReservationGuarantee: null; //unknown
  cpuLimit: null; //unknown
  memoryLimit: null; //unknown
  cpuShares: null; //unknown
  memoryShares: null; //unknown
  extraConfigs: null; //unknown
  namedVmGroups: null; //unknown
  pvdcNamedVmGroupsMap: null; //unknown
  logicalVmGroupReferences: null; //unknown
  pvdcLogicalVmGroupsMap: null; //unknown
  isVgpuPolicy: boolean;
}